[MNT] Restructure the `trend` forecasters module (#5242)
Restructure the trend.py into a trend folder since the file is quiet
large and will get even larger if the FunctionForecaster is added.

See discussion in #5240
