+++ sktime/forecasting/trend/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for trend forecasters."""
