+++ sktime/transformations/series/detrend/tests/test_detrend.py
@@ -4,8 +4,8 @@ import pandas as pd
 import pytest
 
 from sktime.datasets import load_airline
-from sktime.forecasting.tests.test_trend import get_expected_polynomial_coefs
 from sktime.forecasting.trend import PolynomialTrendForecaster
+from sktime.forecasting.trend.tests.test_trend import get_expected_polynomial_coefs
 from sktime.transformations.series.detrend import Detrender
 
 __author__ = ["mloning", "KishManani"]
